+++ app/Http/Controllers/SubCommentController.php
@@ -2,30 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Role;
 use App\Models\SubComment;
 use Illuminate\Http\Request;
 
 class SubCommentController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created subcomment in storage.
-     */
     public function store(Request $request)
     {
         $formFields = $request->validate([
@@ -40,9 +22,7 @@ class SubCommentController extends Controller
         return redirect()->back()->with('message', 'New reply created');
     }
 
-    /**
-     * Store flag on the review.
-     */
+    // Store flag on the review.
     public function flag(string $id)
     {
         SubComment::where('id', $id)->update(['is_flagged' => "1"]);
@@ -51,11 +31,19 @@ class SubCommentController extends Controller
     // Confirm flagg removal
     public function confirm_flag(SubComment $subcomment)
     {
+        abort_if(
+            (auth()->user()->role_id != Role::IS_MODERATOR) &&
+            (auth()->user()->role_id != Role::IS_ADMIN), 403, 'Page doesnt exist' );
+            
        return view('subcomments.remove-flag', ['subcomment' => $subcomment]);
     }
     // Remove flagg
     public function remove_flag(Subcomment $subcomment)
     {
+
+        abort_if(
+            (auth()->user()->role_id != Role::IS_MODERATOR) &&
+            (auth()->user()->role_id != Role::IS_ADMIN), 403, 'Page doesnt exist' );
         $subcomment->where('id', $subcomment->id)->update(['is_flagged' => false]);
         return redirect('reviews')->with('message', 'Flagg removed from subcomment');
     }
@@ -63,49 +51,35 @@ class SubCommentController extends Controller
 
     // Get confirm-hide view
     public function confirm_hide(Subcomment $subcomment){
+        abort_if(
+            (auth()->user()->role_id != Role::IS_MODERATOR) &&
+            (auth()->user()->role_id != Role::IS_ADMIN), 403, 'Page doesnt exist' );
+
         return view('subcomments.hide', ['subcomment' => $subcomment]);
     }
-        /**
-     * Hide subcomment.
-     */
+        //Hide subcomment.
     public function hide(Subcomment $subcomment)
     {
+        abort_if(
+            (auth()->user()->role_id != Role::IS_MODERATOR) &&
+            (auth()->user()->role_id != Role::IS_ADMIN), 403, 'Page doesnt exist' );
+
         $subcomment->where('id', $subcomment->id)->update(['is_deleted' => true]);
         return redirect('/reviews');
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id)
-    {
-        //
     }
     
     // Show confirm-delete view
     public function confirm_delete(SubComment $subcomment)
     {
+        abort_if(auth()->user()->role_id != Role::IS_ADMIN, 403, 'Page doesnt exist');
+
         return view('/subcomments/delete', ['subcomment' => $subcomment]);
     }
     //Remove the specified comment from storage.
     public function destroy(SubComment $subcomment)
     {
+        abort_if(auth()->user()->role_id != Role::IS_ADMIN, 403, 'Page doesnt exist');
+        
         $subcomment->delete();
         return redirect('/reviews')->with('message', 'Reply succesfully removed');
     }
